requirements.txt
numpy
sounddevice
scipy
pySerial
keyboard